requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
python-barcode>=0.15.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
werkzeug>=3.0.0
